# Use shared pre-computation for `by` and `perm` orderings.