--- a/base/ordering.jl
+++ b/base/ordering.jl
@@ -193,18 +193,18 @@ lt(o::ForwardOrdering, a, b) = isless(a, b)
 # Reverse
 prepare(o::ReverseOrdering, x) = prepare(o.fwd, x)
 lt_prepared(o::ReverseOrdering, a, b) = lt_prepared(o.fwd, b, a)
-lt(::ReverseOrdering, a, b) = lt_prepared(o, prepare(o, a), prepare(o, b))
+lt(o::ReverseOrdering, a, b) = lt_prepared(o, prepare(o, a), prepare(o, b))
 
 # By
 prepare(o::By, x) = prepare(o.order, o.by(x))
 lt_prepared(o::By, a, b) = lt_prepared(o.order, a, b)
-lt(::By, a, b) = lt_prepared(o, prepare(o, a), prepare(o, b))
+lt(o::By, a, b) = lt_prepared(o, prepare(o, a), prepare(o, b))
 
 # Perm
 @propagate_inbounds prepare(o::Perm, i) = (prepare(o.order, o.data[i]), i)
 lt_prepared(p::Perm, (da, a), (db, b)) =
     (lt_prepared(p.order, da, db)::Bool) | (!(lt_prepared(p.order, db, da)::Bool) & (a < b))
-@propagate_inbounds lt(::Perm, a, b) = lt_prepared(o, prepare(o, a), prepare(o, b))
+@propagate_inbounds lt(o::Perm, a, b) = lt_prepared(o, prepare(o, a), prepare(o, b))
 
 ## Lt
 lt(o::Lt, a, b) = o.lt(a, b)