- Add internal order preparation API to let sorting algorithms prepare elements for repeated comparisons
- Use the API
- Add performance tests

How many times `by` is called
```julia
cnt = Ref(0)
incr_identity = x -> (cnt[] += 1; x)
x = 1:50

cnt[] = 0
@test issorted(x; by=incr_identity)
println(cnt[]) # 98 => 50

cnt[] = 0
@test searchsortedfirst(x, 1; by=incr_identity) == 1
printlnt(cnt[]) # 17 => 7

cnt[] = 0
@test searchsorted(repeat(1:10, inner=10), 3; by=incr_identity) == 21:30
println(cnt[]) # 26 => 16

cnt[] = 0
x = hash.(1:1000) # stable rng lol
@test sort(x; by=incr_identity) == sort(x)
println(cnt[]) # 24638 => 12999
```

Runtime performance tests
```julia
by_f = x -> exp(x) % 3
x = 1:50
sx = sort(x; by=by_f);
y = hash.(1:1000); # stable rng lol

@btime issorted(x; by=$by_f) # 30.066 ns => 30.570 ns
@btime issorted(sx; by=$by_f) # 977.417 ns => 471.649 ns
@btime searchsortedfirst($sx, 1; by=$by_f) # 161.942 ns => 129.231 ns
@btime searchsorted($sx, 3; by=$by_f) # 176.821 ns => 132.665 ns
@btime sort($y; by=$by_f); # 8.569 μs (3 allocations: 7.88 KiB) => 7.385 μs (3 allocations: 7.88 KiB)
```

Something to maybe consider doing later is marking this `lt_prepared` API public.

A downside of this PR is that it makes comparing with an `Ordering` subtype that defines neither `lt` nor `lt_prepared` a stack overflow error instead of a method not found error. This could happen of someone defines `MyOrder <: Ordering` and `lt(::MyOrder, ::MyType, ::MyType)` and then calls `lt(o::MyOrder, 1, 2)`. It's possible to remove that downside by replacing
```julia
lt(::Ordering, x, y) = lt_prepared(o, prepare(o, x), prepare(o, y))
```
with
```julia
lt(::Union{ReverseOrdering, By, Perm}, x, y) = lt_prepared(o, prepare(o, x), prepare(o, y))
```
But that makes the (internal) `lt_prepared` API more cludgy to use.