@nsajko 2023-11-08
Fabulous idea!

> A downside of this PR is that it makes comparing [...]

Couldn't you fix that by using "traits" to declare that a type supports the new API?

```julia
struct Preparable end
struct NotPreparable end

preparable(::Ordering) = NotPreparable()

preparable(::ReverseOrdering) = Preparable()
preparable(::By) = Preparable()
preparable(::Perm) = Preparable()
```

If you need `preparable` to be recursive (didn't look closely), you could, e.g., replace `preparable(::Perm) = Preparable()` with `preparable(p::Perm) = preparable(p.order)`.