@LilithHafner 2023-11-08
We could do that, or we could say the "trait" is to define 
```julia
lt(::MyNewPreparableOrdering, x, y) = lt_prepared(o, prepare(o, x), prepare(o, y))
```
In order of verbosity:
1) PR (stack overflow)
2) This comment's "trait"s
3) Full traits

Full traits also allow algorithmic dispatch based on the presence of preparability, but I think that is actually something we do not want to support because I want supporting preparation to remain totally free (i.e. never a regression vs not supporting preparation). If you have a comparison that is 90% lt and 10% preparation, it still makes sense to share the preparation and I don't want to dispatch based on the existence of preparation for that reason.

Unfortunately, preparability is only recursive if each Ordering supports it explicitly.

`MyUnpreparableBy(-, By(abs))` is not preferable because there is no way (that I know of) for us to figure out what the appropriate input to `abs` is without the cooperation of `MyUnpreparableBy`'s author.