@LilithHafner 2023-11-22
The API this adds is internal, the downside in the OP is no longer present, and I'd like to merge this before working on sortperm performance so that that stuff cleanly generalizes to by orders as well. So I'm inclined to merge this unless anyone objects.